extracted createCycleMiddleware and combineCycles
